Move EventSubscription into EventSubscriptionHandlers
